% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bing.R
\name{bing}
\alias{bing}
\title{Search Bing}
\usage{
bing(search_terms)
}
\arguments{
\item{search_terms}{Search terms encapsulated in " ".}
}
\description{
Improve your workflow by searching Bing directly from the console without having to switching to the browser and
opening a new tab first.
}
\examples{
bing("my search terms")
}
\keyword{bing}
\keyword{internet}
\keyword{workflow}
