% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossref.R
\name{crossref}
\alias{crossref}
\title{Find DOI's and other bibliographic metadata}
\usage{
crossref(search.terms)
}
\arguments{
\item{search.terms}{Search terms encapsulated in " ".}
}
\description{
Search crossref.org directly from the R console without having to switching to the browser and
opening a new tab first.
}
\examples{
crossref("my source")
}
\keyword{DOI}
\keyword{bibliography}
\keyword{internet}
\keyword{workflow}
