% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{github}
\alias{github}
\title{Search GitHub}
\usage{
github(search_terms)
}
\arguments{
\item{search_terms}{Search terms encapsulated in " ".}
}
\description{
Improve your workflow by searching GitHub directly from R console.
}
\examples{
github("ggplot extensions")
}
\keyword{github}
\keyword{internet}
\keyword{workflow}
