% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google.hits.R
\name{google.hits}
\alias{google.hits}
\title{Obtain number of Google hits}
\usage{
google.hits(search.terms)
}
\arguments{
\item{search.terms}{Search terms encapsulated in "".}
}
\description{
Quickly get a grasp of how much a search term or a combination of search terms are searched for on google.

Inspired by this discussion on Stackoverflow: https://stackoverflow.com/questions/30199157/r-google-search-result-count-retrieve
}
\examples{
google.hits("my search terms")

# Your Search URL:
# https://www.google.com/search?q=my\%20search\%20terms

# No. of Hits:
# [1] 117000000
}
\keyword{google}
\keyword{hits}
\keyword{internet}
