% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google.R
\name{google}
\alias{google}
\title{Search Google}
\usage{
google(search.terms)
}
\arguments{
\item{search.terms}{Search terms encapsulated in " ".}
}
\description{
Improve your workflow by searching google directly from the console without having to switching to the browser and
opening a new tab first.
}
\examples{
google("my search terms")
}
\keyword{google}
\keyword{internet}
\keyword{workflow}
