% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google.scholar.R
\name{google.scholar}
\alias{google.scholar}
\title{Search Google Scholar}
\usage{
google.scholar(search.terms)
}
\arguments{
\item{search.terms}{Search terms encapsulated in " ".}
}
\description{
Improve your workflow by searching Google Scholar directly from the console without having to switching to the browser and
opening a new tab first.
}
\examples{
google.scholar("my search terms")
}
\keyword{google}
\keyword{internet}
\keyword{scholar}
\keyword{workflow}
