% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yandex.R
\name{init_yandex_search}
\alias{init_yandex_search}
\title{Initiate yandex search}
\usage{
init_yandex_search(user_name, api_key, sort_by = c("rlv", "tm"),
  filter = c("none", "moderate", "strict"))
}
\arguments{
\item{user_name}{Yandex user}

\item{api_key}{API key given to Yandex user}

\item{sort_by}{Sort mode: \code{rlv} - relevance, \code{tm} time}

\item{filter}{filter mode: \code{none}, \code{moderate} or \code{strict}}
}
\description{
Function takes more or less stable parameters and returns another
unified search function that takes 3 parameters:
query, numner of results on page and page number
}
