% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors_warnings.r
\name{last_error}
\alias{last_error}
\title{Get last error}
\usage{
last_error()
}
\description{
This is a helper function that returns the last produced error message in R, which is intended to be used within any of the search functions of the {websearchr} package. It adds the keyword "R" to the beginning for improved accuracy of search results.
}
\examples{
# Query last error message
last_error()
}
\keyword{error}
\keyword{message}
