% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors_warnings.r
\name{last_warning}
\alias{last_warning}
\title{Get last warning}
\usage{
last_warning()
}
\description{
This is a helper function that returns the last produced warning message in R, which is intended to be used within any of the search functions of the {websearchr} package. It adds the keyword "R" to the beginning for improved accuracy of search results.
}
\examples{
\dontrun{
# Create example warning
warning("Example warning")

# Query last warning message
last_warning()
}
}
\keyword{message}
\keyword{warning}
