% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_bloggers.R
\name{r_bloggers}
\alias{r_bloggers}
\title{Search r-bloggers.com}
\usage{
r_bloggers(search_terms)
}
\arguments{
\item{search_terms}{Search terms encapsulated in " ".}
}
\description{
Improve your workflow by searching r-bloggers.com directly from the console without having to switching to the browser and
opening a new tab first.
}
\examples{
r_bloggers("my search terms")
}
\keyword{internet}
\keyword{r-bloggers}
\keyword{workflow}
