% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdocumentation.R
\name{rdocumentation}
\alias{rdocumentation}
\alias{rdoc}
\title{Search rdocumentation.org}
\usage{
rdocumentation(search.terms)

rdoc(search.terms)
}
\arguments{
\item{search.terms}{Search terms encapsulated in " ".}
}
\description{
Improve your workflow by searching rdocumentation.org directly from the console without having to switching to the browser and
opening a new tab first.
}
\examples{
rdocumentation("my search terms")
rdoc("package I am looking for")
}
\keyword{internet}
\keyword{rdocumentation}
\keyword{workflow}
