% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdrr.io.R
\name{rdrr.io}
\alias{rdrr.io}
\alias{rdrr}
\title{Search rdrr.io}
\usage{
rdrr.io(search.terms)

rdrr(search.terms)
}
\arguments{
\item{search.terms}{Search terms encapsulated in " ".}
}
\description{
Improve your workflow by searching rdrr.io directly from the console without having to switching to the browser and
opening a new tab first.
}
\examples{
rdrr.io("my search terms")
rdrr("my search terms")
}
\keyword{internet}
\keyword{rdrr.io}
\keyword{workflow}
