% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdrr_io.R
\name{rdrr_io}
\alias{rdrr_io}
\alias{rdrr}
\title{Search rdrr.io}
\usage{
rdrr_io(search_terms)

rdrr(search_terms)
}
\arguments{
\item{search_terms}{Search terms encapsulated in " ".}
}
\description{
Improve your workflow by searching rdrr.io directly from the console without having to switching to the browser and
opening a new tab first.
}
\examples{
rdrr_io("my search terms")
rdrr("my search terms")
}
\keyword{internet}
\keyword{rdrr.io}
\keyword{workflow}
