% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reddit.r
\name{reddit}
\alias{reddit}
\title{Search Reddit}
\usage{
reddit(search_terms)
}
\arguments{
\item{search_terms}{Search terms encapsulated in " ".}
}
\description{
Improve your workflow by searching Reddit directly from the console without having to switching to the browser and
opening a new tab first.
}
\examples{
reddit("my search terms")
}
\keyword{internet}
\keyword{reddit}
\keyword{workflow}
