% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startpage.R
\name{startpage}
\alias{startpage}
\alias{sp}
\title{Search Startpage}
\usage{
startpage(search_terms)

sp(search_terms)
}
\arguments{
\item{search_terms}{Search terms encapsulated in " ".}
}
\description{
Improve your workflow by searching with startpage.com directly from the console without having to switching to the browser and opening a new tab first.
}
\examples{
startpage("my search terms")
sp("r-project")
}
\keyword{internet}
\keyword{startpage}
\keyword{workflow}
