% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twitter.R
\name{twitter}
\alias{twitter}
\title{Search Twitter}
\usage{
twitter(search_terms, lang = c("en", "de", "es", "fr"))
}
\arguments{
\item{search_terms}{Search terms encapsulated in " ".}

\item{lang}{Specify in which language Twitter should be accessed One of c("en", "de", "es" "fr") for English, German, Spanish and French, respectively.}
}
\description{
Improve your workflow by searching Twitter directly from the console without having to switching to the browser and
opening a new tab first.
}
\examples{
twitter("rstudiotips")
}
\keyword{internet}
\keyword{wikipedia}
\keyword{workflow}
