% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web.R
\name{web}
\alias{web}
\alias{open}
\title{Access Domain}
\usage{
web(address, https = TRUE, suppressWWW = FALSE)

open(address, https = TRUE, suppressWWW = FALSE)
}
\arguments{
\item{address}{Open a web address. No quotation marks required!}

\item{https}{if FALSE "http" will be used instead of the default "https".}

\item{suppressWWW}{if TRUE "www" will be suppressed and the user input will follow directly after https://}
}
\description{
Improve your workflow by accessing web directly from R console.
}
\examples{
web(r-project.org) # no quotation marks 
open(github.com) # no quotation marks 
}
\keyword{web}
\keyword{workflow}
