% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikipedia.R
\name{wikipedia}
\alias{wikipedia}
\alias{wp}
\title{Search Wikipedia}
\usage{
wikipedia(search_terms, lang = c("en", "de", "es", "fr"))

wp(search_terms, lang = c("en", "de", "es", "fr"))
}
\arguments{
\item{search_terms}{Search terms encapsulated in " ".}

\item{lang}{In which language Wikipedia should be accessed. One of c("en", "de", "es" "fr") for English, German, Spanish and French, respectively.}
}
\description{
Improve your workflow by searching Wikipedia directly from the console without having to switching to the browser and
opening a new tab first.
}
\examples{
wikipedia("my search terms")
wp("my search terms")
}
\keyword{internet}
\keyword{wikipedia}
\keyword{workflow}
