% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wolframalpha.r
\name{wolfram}
\alias{wolfram}
\title{Search WolframAlpha}
\usage{
wolfram(search_terms)
}
\arguments{
\item{search_terms}{Search terms encapsulated in " ".}
}
\description{
Improve your workflow by searching WolframAlpha directly from the console without having to switching to the browser and
opening a new tab first.
}
\examples{
wolfram("my search terms")
}
\keyword{internet}
\keyword{wolframalpha}
\keyword{workflow}
