% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yandex.R
\name{yandex_error}
\alias{yandex_error}
\title{Describe error using code}
\usage{
yandex_error(code)
}
\description{
Documentation is available on \link{https://tech.yandex.com/xml/doc/dg/concepts/about-docpage/}
}
