% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yandex.R
\name{yandex_group_by}
\alias{yandex_group_by}
\title{Create group_by parameter value for init_yandex_search function}
\usage{
yandex_group_by(mode = c("flat", "deep"), attr = c(" ", "d"),
  groups_on_page = 10, docs_in_group = 1)
}
\arguments{
\item{mode}{Grouping method. Possible values:
\code{flat} — Flat grouping. Each group contains a single document.
\code{deep} — Grouping by domain. Each group contains documents from a single domain.}

\item{attr}{Utility attribute. Depends on the value of the mode attribute.}

\item{groups_on_page}{Maximum number of groups that can be returned per page of search results.
Acceptable values - from 1 to 100.}

\item{docs_in_group}{— Maximum number of documents that can be returned per group.
Acceptable values — from 1 to 3.}
}
\description{
please see Yandex.XML documentation \link{https://tech.yandex.com/xml/doc/dg/concepts/about-docpage/}
}
