% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yandex.R
\name{yandex_search}
\alias{yandex_search}
\title{Low-level yandex search function}
\usage{
yandex_search(user, key, query, lr, l10n, sortby, filter, maxpassages,
  groupby, page, showmecaptcha)
}
\description{
Low-level yandex search function
}
