% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckduckgo.R
\name{duckduckgo}
\alias{duckduckgo}
\alias{ddg}
\title{Search Duckduckgo}
\usage{
duckduckgo(search.terms)

ddg(search.terms)
}
\arguments{
\item{search.terms}{Search terms encapsulated in " ".}
}
\description{
Improve your workflow by searching with duckduckgo.com directly from the console without having to switching to the browser and
opening a new tab first.
}
\examples{
duckduckgo("my search terms")
ddg("r-project")
}
\keyword{duckduckgo}
\keyword{internet}
\keyword{workflow}
