% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qwant.R
\name{qwant}
\alias{qwant}
\title{Search Qwant}
\usage{
qwant(search.terms, lang = c("en", "de", "es", "fr"))
}
\arguments{
\item{search.terms}{Search terms encapsulated in " ".}

\item{lang}{In which language Qwant should be accessed. One of c("en", "de", "es" "fr") for English, German, Spanish and French, respectively.}
}
\description{
Improve your workflow by searching Qwant directly from the console without having to switching to the browser and
opening a new tab first.
}
\examples{
qwant("my search terms")
}
\keyword{internet}
\keyword{qwant}
\keyword{workflow}
