% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qwant.R
\name{qwant}
\alias{qwant}
\title{Search Qwant}
\usage{
qwant(search_terms)
}
\arguments{
\item{search_terms}{Search terms encapsulated in " ".}
}
\description{
Improve your workflow by searching Qwant directly from the console without having to switching to the browser and
opening a new tab first.
}
\examples{
qwant("my search terms")
}
\keyword{internet}
\keyword{qwant}
\keyword{workflow}
