% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stackoverflow.R
\name{stackoverflow}
\alias{stackoverflow}
\alias{so}
\title{Search Stackoverflow}
\usage{
stackoverflow(search.terms)

so(search.terms)
}
\arguments{
\item{search.terms}{Search terms encapsulated in " ".}
}
\description{
Improve your workflow by searching Stackoverflow directly from R console.
}
\examples{
stackoverflow("r date conversion")
so("r ggplot2 geom_smooth()")
}
\keyword{stackoverflow}
\keyword{web}
\keyword{workflow}
