% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twitter.R
\name{twitter}
\alias{twitter}
\title{Search Twitter}
\usage{
twitter(search.terms)
}
\arguments{
\item{search.terms}{Search terms encapsulated in " ".}
}
\description{
Improve your workflow by searching Twitter directly from the console without having to switching to the browser and
opening a new tab first.
}
\examples{
twitter("@rstudiotips")
}
\keyword{internet}
\keyword{wikipedia}
\keyword{workflow}
