% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikipedia.R
\name{wikipedia}
\alias{wikipedia}
\alias{wp}
\title{Search Wikipedia}
\usage{
wikipedia(search.terms)

wp(search.terms)
}
\arguments{
\item{search.terms}{Search terms encapsulated in " ".}
}
\description{
Improve your workflow by searching Wikipedia directly from the console without having to switching to the browser and
opening a new tab first.
}
\examples{
wikipedia("my search terms")
wp("my search terms")
}
\keyword{internet}
\keyword{wikipedia}
\keyword{workflow}
